+++ src/components/BurgerIngredients/BurgerIngredients.module.css
@@ -2,30 +2,44 @@
   max-width: 600px;
 }
 
-.totalPrice {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.totalPrice > svg {
-  width: 33px;
-  height: 33px;
-}
-
-.scrollBox {
+.ingredientsBox {
+  height: 75%;
   overflow: auto;
-  max-height: 35vh;
 }
 
-.scrollBox::-webkit-scrollbar {
+.ingredientsBox::-webkit-scrollbar {
   width: 8px;
 }
-.scrollBox::-webkit-scrollbar-track {
+.ingredientsBox::-webkit-scrollbar-track {
   background: #2F2F37; 
 }
-.scrollBox::-webkit-scrollbar-thumb {
+.ingredientsBox::-webkit-scrollbar-thumb {
   background-color: #8585AD;
   border-radius: 20px;  
   border: 3px; 
+}
+
+.ingredients {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+}
+
+.ingredientsCard {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.ingredientsCard:nth-child(n+3) {
+  margin-top: 32px;
+}
+
+.price {
+  display: flex;
+}
+
+.cardName {
+  max-width: 250px;
+  text-align: center;
 }
